cursor-class Create new parent Cursor class and refactor Velocity to extend it
